Set some device limits + fix warnings

+++ src/engine.cpp
@@ -27,7 +27,7 @@ bool Engine::get_use_mirror_window()
 #endif
 }
 
-void Engine::update(double delta_time)
+void Engine::update(float delta_time)
 {
     renderer->update(delta_time);
 }
